Add "typecho-core" engine and "@fancyapps/ui" into whitelist.
As you saw the former package is used to act as a new element-based engine for Typecho-based projects, which provides a better SPA experience for its web users; and the later "@fancyapps" is the developer organization of Fancybox. jQuery fancybox ended its life on version 3 and this package is a ES6-based generation of a same project, providing gallery & thumbs utilities.